fix mouse scroll in fixedlist
There used to be a comment in GUIFixedListContainer saying

// calling SelectItem() here will focus the item and scroll, which
// isn't really what we're after

however, the subsequent dirty check triggered through DoProcess ->
GUIControlGroup::Process -> UpdateVisibility would trigger a
SelectItem nevertheless and render this old hack ineffective! Causing
the erratic and unusable scroll behaviour of the fixedlist when
hovering with the mouse.

To stop this, we introduce a new private flag that gets set whenever
the mouse is moved. This will tell the upcoming call to SelectItem to
ignore the fixed position scrolling and thus restore the original
mouse scrolling code in GUIFixedListContainer::SelectItemFromPoint
(which was broken due to this dirty refresh).