[filesystem] fix credentials for smb protocol in case of vfs add-on
currently we hide credentials for smb urls and store them into PasswordManager. Internal smb implementation knows about this and properly gets credentials from the manager and operates with full url but a VFS add-on can't do this. To fix this I've move this logic to lower level to CDirectory and CFile. This allows smb VFS add-on work properly.