guilib/GUIWindow: remove obsolete GUIEvent declarations
The macros `ON_CLICK_MESSAGE` and `ON_SELECTED_MESSAGE` are not used
at all.

The attributes `m_mapClickEvents` and `m_mapSelectedEvents` are read
in `OnMessage()` (in a very clumsy way which auto-creates empty map
items), but are never written to.

The constant `GUI_MSG_SELCHANGED` is checked in `OnMessage()`, but is
never sent.

The `GUICallback.h` library is not used by anybody anymore, so this
commit removes it completely.