guilib/MatrixGLES: remove implicit zero-initialization
This made the class non-trivial and that alone added runtime overhead,
because it disabled compiler optimizations.  Additinally, an all-zero
matrix is useless, and is therefore not a good default value.  Without
"real" initialization, it is more dangerous to use it, because that
hides bugs (uses of uninitialized data) from Valgrind.

So instead of doing bogus initialization, don't initialize it at all,
and let Valgrind show where the real bugs are.