FileItem: determine mimetype based on DynPath
Since commit 22b1e1e fileitems within playlists
hold the filename of the playlist in m_strPath.
Therefore we need to ask m_strDynPath for the mimetime, otherwise
we end up trying to expand the item as playlist.

For a live http video stream we read endless amount of data
and add garbage to a playlist which is none.