VideoPlayer: fix http buffering for live http streams
We observe ocassional buffering shortly after starting http
live streams.
When the fileitem's path is a pvr item and the dynpath is http
(this is the case when using pvr.iptvsimple for example) we assume
that the stream is a livestream and let DVDInputStreamFFmpeg handle it.
Additionally we flag it as a realtime stream.