windowing/gbm: plane selection rework
This rework now allows us to find a dedicated plane for video
and a dedicated plane for the gui. Before we would swap the gui
back and forth between planes depending if a video was playing.
That behaviour isn't needed and now we can run the gui on the overlay
full time.

We have to track the fb format and use an alpha format when a
video is playing and a non-alpha format when no video is playing.
When the format changes the fb is validated and destroyed so a new
fb can be created.

This also allows us to add dependencies to the primary plane. For example
in the DRMPRIME renderer we can disable it if no primary plane is available.

We check for three formats currently:
1) YUV420 -> for RPi
2) YUYV   -> for i.MX6
3) NV12   -> for basically everything else

more formats can be added in the future if needed.