[database] Reset path hashes when cleaning up files in video database
By performing a video database clean-up, effectively the state in the
video database is brought in sync with the state of the source in terms
of files that were removed from the source. This has to be reflected by
updating the hash of the path so the next library scan is not operating
on outdated information. However, since added items are not reflected,
it would be wrong to update the path hash to the actual directory
contents. Just resetting the hash for any path that had changes (i.e. saw
removed files) makes sure that the next library update does not miss both
newly added files and files that were removed before a clean-up but later
re-added.