Customize Estuary Home Screen and Widgets
This modification to Estuary provides new options to:
1- have home screen menu items reordered
2- have home menu item widgets disabled
3- have category widgets show as small icons
3- show a TV Show widget the Unwatched/In Progress TV Shows

By default these mods provide the exact same setup as Estuary does currently, including game support. The one thing that is less intuitive is if you have a home section with no content, the DISABLE THIS MENU ITEM button goes to the skin settings instead of just disabling the option. The old way of disabling wouldn't work because of the option to reorder home menu items, and I couldn't find a way for the button to take you right to the menu options skin settings section.

There is currently no "upgrade" from the old settings, so Estuary would reset to its default.  I have thought of a way to do a one time settings read of home menu items, but it's a fair amount of work that I'd rather not do unless the PR is accepted.

These modifications require no additional addons.