Game API v1.1.0: Add feature names to controller layouts
This change expands the existing controller layout with vectors of
names for the various types of features on each controller.

The API has been modified to report controller layouts using a dedicated
function, which is called after the input topology is loaded. Other API
calls have been simplified to just pass controller ID instead of the full
layout.