[music] Support originaldate / originalyear tags
The MusicBrainz Picard application places the release date of an album
in its "date" tag. But most users expect to see the *original* release
date of an album, which Picard places in its "originaldate" tag.

See: https://picard.musicbrainz.org/docs/mappings/

For example, a 1994 re-release of Led Zeppelin IV:

  $ metaflac --export-tags-to - "1 - Black Dog.flac" | sort
  ...
  DATE=1994-07-19
  ...
  ORIGINALDATE=1971-11-08
  ORIGINALYEAR=1971

Parse various tag formats for the *original* release date or year, and
prefer that over the release date in the music database.

To see the changes take effect, users will need to touch all music
files and then rebuild the music library.