Max 60s commbreakautowait and commbreakautowind
Some users need to set my commbreakautowait to a value above 10 (30 in my case).
Here is a link to a thread I found while searching for why commbreakautowait was not working.
From looking at the code, then testing, I can see that it does work, but is limited to 10 seconds.
http://forum.kodi.tv/showthread.php?tid=257520

Update autowind and autowait logic

Update the autowait and autowind logic to handle the extended times.
This change will in effect cause a cut to be skipped when the autowait + autowind values are greater than the length of the cut, but setting the start and end of the cut to the same time.
The code will not allow the end time for the cut to be less than the start time.

Update xbmc/settings/AdvancedSettings.cpp

Allow negative comm skip time for increased flexibility to support intricacies of comm flagging with various backends.

Co-Authored-By: Andrew Fowler <metaron-uk@users.noreply.github.com>

Update xbmc/settings/AdvancedSettings.cpp

Allow negative comm skip time to allow for backends that are late at flagging commercial breaks.

Co-Authored-By: Andrew Fowler <metaron-uk@users.noreply.github.com>