[python] Check version of cryptography module
We have had multiple issues with Ubuntu 16.04 using an outdated version
of the Python cryptography module that causes strange SSL behavior.
Example for this is not being able to connect to SSL sites.

This commit introduces a check on initialization of the interpreter
to make sure that we use a compatible version.

See also:
https://github.com/pyca/pyopenssl/issues/542#issuecomment-312968275
https://forum.kodi.tv/showthread.php?tid=335786