[PVR] Ensure PVR_MENUHOOK dialog is always opened, even for one item.
The PVR_MENUHOOK allows a PVR client to create a "client actions"
sub-menu in the PVR system (timers, epg, etc.)

Previously, if there is only one option in the sub-menu then it is
automatically selected and actioned. So, if the sub-menu has the
option "delete forever" and the user presses "context menu" then
selects "client actions", then the recording would be deleted without
any user action. This can be demonstrated with "pvr.demo".

Now, we always open the selection dialog to allow the user to see and
explicitly select the option or press the existing "cancel" button on
the dialog.