AEStreamParser: Fix E-AC-3 dependent stream passthrough (trac#17871)
The current handling of E-AC-3 dependent streams in AEStreamParser has
several issues:

- AddData() returns no data on the main AC-3 frame. This happens even if
  it was provided a combined main frame and a dependent frame, such as
  those provided by FFmpeg 4.0+ after ae9297097696f3 ("avcodec/eac3: add
  support for dependent stream").
  (For the record, the main frame and dependent frame provide audio data
  for the same timestamp so it makes sense to consider them as a single
  frame.)

  This is problematic as AddData() not returning a frame is interpreted
  to mean that it needs more data, causing the caller to call it again
  with more data. This repeats with each frame, causing the caller input
  buffer to be extended with more and more unneeded data over time (i.e.
  it is called twice for each combined frame).

  This also causes increasing timestamp drift as the caller
  (CDVDAudioCodecPassthrough) does not expect there to be multiple frames
  in the data backlog and it uses the dts of new incoming frames (that
  are put at the end of the backlog) as the output dts of frames received
  from AEStreamParser::AddData() which is actually significantly behind.

  This causes trac#17871.

- The current concatenation code does not properly handle the case where
  resyncing occurs between the main frame and dependent frame. In such a
  case garbage may be output.

- The current code in SyncAC3() does not properly validate that there is
  enough data in the buffer in the (m_fsizeMain != 0) case, allowing
  invalid input to cause a crash.

Fix the issues by reworking the E-AC-3 dependent stream support so that
when we do not yet know if the stream is AC-3 or E-AC-3 or we know
that the stream is E-AC-3 and an AC-3 (main) frame is encountered,
look if an E-AC-3 dependent frame immediately follows.

If so, return a concatenated frame without requiring multiple AddData()
calls. If not, consider the stream as AC-3. Do not try to skip over
garbage bytes, desync/resync will be handled by the next call if
required.

If we want to perform the above lookup but there is not enough data in
the input buffer, use the m_needBytes mechanism to request more data,
similar to how the DTS HD code works. This ensures that on return from
AddData() either a valid frame is returned or more data is needed.

Fixes trac#17871.