Fix infinite loop when listitem action is "Play"
This fixes an infinite loop, noticed in the Game OSD, when clicking the
"Play" item. Actions are sent to the GUI before the player, and the GUI
(CGUIBaseContainer) considers the "Play" action to be a "Select" action.
It then resolves the "Select" action into "Play", which causes an infinite
loop.