[joystick.xml] Add more buttons to bring up OSD and Menu in EPG
This makes the <x> and <back> buttons open the OSD in FullscreenTV and FullscreenRadio, like they already do in FullscreenVideo. Also the <back> button can now be used in TVGuide/RadioGuide to open the Menu. Without this, there's no way to open it and <back> didn't have any function before, anyway.