[wayland] Centralize seat handling
Previously, Wayland seat (including input) handling was performed for
the main surface and the window decorations in separate places,
leading to multiple instances of wl_seat, wl_pointer etc.

While not forbidden by the protocol, this has caused aberrant behavior
in compositors such as KWin which were then not able to hide the cursor
image. As it is indeed unclear what should happen with the cursor when
multiple wl_pointer instances are created, the solution of the problem
is to centralize the seat handling on the Kodi side.
There is now only one wl_seat etc. instance that retrieves the events
from the Wayland compositor and distributes them further in the Kodi
application to the relevant components (generating Kodi events or
handling decorations).

Fixes #15002