ActiveAE: Don't return null stream when AE is suspended
If you start playback of an audio file/playlist while
a refreshrate switch is happening, we skip tracks or playback
fails at all.

The IsSuspended check is not necessary, as the MakeStream request
has a timeout of 10s anyway.
So let it fail if the AudioEngine is supsended for more than 10s.