[addons] Do not list non-repo add-ons as "unavailable" in info dialog
Current GUIDialogAddonInfo logic only displays a detailed entry in the
add-on dependency list if the add-on in question can be found in a
repository. Add-ons that are only available locally (pre-shipped with
Kodi or by distributor) such as script.module.pil in the default
installation are wrongly listed as "unavailable" even though they are
correctly found and enabled (if necessary) when installing the add-on.

In fact a locally available version is already looked up, but
completely disregarded if no repository version is found. This commit
fixes that situation by falling back to local add-on data if none
is found in the repositories.

Fixes #14697