[guilib] Fix out-of-bounds vector access when "drawing" empty text
If the text to draw is empty, the vertex vector in CGUIFontTTFBase::
DrawTextInternal stays empty and is handed like that to the lib-specific
functions which might not handle that. Specifically, CGUIFontTTFGL::
CreateVertexBuffer accessed the vertex vector without bounds checking
(detected by `-D_GLIBCXX_ASSERTIONS`). As there is nothing to do for
drawing an empty text anyway, just bail out early.
Also add some extra checking to be safe in the future.