[addons] Workaround file copying bug decoding "+" to space in file names
Installation of add-ons with "+" in their name or version currently
fails because the downloaded file cannot be found. This is due to path
decoding applied in the course of calling CFileOperationJob for
performing the download which replaces "+" with space.

Actually, decoding "+" to space is an error for URLs (only to be
performed for the query part and not other parts), but changing this
in the RC phase could lead to too many regressions in other parts of the
application.

There is also a design flaw with CFileOperationJob in general because
(to my knowledge) it does not tell you the resulting download location,
so without using special functions it is always possible that the final
file name is not what the calling code expects.

Fixes #14779