Update joystick.xml
Added <back>ActivateWindow(Home)</back> as a global keypress, it was not assigned any action before.
Changed <b> back to the original action of Stop and changed the long press action to Fullscreen.
Changed the <back> button to the Fullscreen action. It doesn't toggle, instead it sends you back to the home screen after pressing it a second time because of the global action, which kind of makes sense for the View (on xbox one, previously back) button.