Active Skin Update - Force Close TextureBundle
Updating the active skin fails due to Kodi maintaing a lock on the  textures.xbt file.

This PR allows the active skin to be updated when installing updated skin via addons "Install from zip file". Tested with default textures and also with themes selected. Not tested when updating skin from online source which may still fail due to antivirus locking the downloaded folder.

This is essentially a copy of PR #11090 which was merged to Krypton but seems to have been omitted in Leia.