Input: Fix ignoring custom joystick.xml
Currently, custom joystick.xml is loaded last, and its buttons are
appended to the existing keymap. However, the keymapper uses the first
action in the keymap, so appended actions may be ignored.

The fix is to clear the keymap for a window/joystick combination when a
new node is discovered in a later keymap XML.