[guilib] Handle empty vertex buffers in GUIFontTTFGL
CGUIFontTTF::DrawTextInternal can end up with zero-element vertex
buffers in corner cases such as when all text to render is completely
clipped away. This should also be cached as such, so it is unfeasible to
forbid empty vertex buffers completely. CGUIFontTTFDX already handles
this case, but CGUIFontTTFGL did not.
A previous crash fix added an assertion on the vertex list not being
empty, but this is not needed anymore since that case is correctly
handled now.