Update PVRTimers.cpp
Fix for this issue:

In the EPG, selecting 'Delete timer' on a program with a series timer rule opens the 'Confirm delete' dialog asking 'Do you only want to delete this timer or also the timer rule that has scheduled it?' with two options 'All' or 'Only this'.

Both options have the same result and only delete the selected timer. The 'All' option fails to delete the timer rule.