CUtil::ValidatePath fix stripping double (back)slash chars
Only used by Python call xbmc.validatePath(). When double slash is found in supplied path, the remaining of the path is stripped of instead of just the double slash. This used to work as intended before the previous change, until Delete(x) got replaced by erase(x). Delete(x) was a function which called erase(x, 1).