Add Permission To AndroidManifest.xml.in
Android 8.0 (API level 26) Brought Changes To Installing APK's

android.permission.REQUEST_INSTALL_PACKAGES is now required if you want to install APK's from inside Kodi. I do not doubt that you know this so maybe the team is choosing to leave it out? Also apks must be provided by content provider thux the apk has to be in the apps Private storage. None of which are issues once the permission is added to the Manifest. Already tested and working.