[curl] Move libcurl symbols out of namespace
Putting includes (that can in turn include all sorts of other stuff)
into namespaces is not really maintainable and finally broke the build
in Linux 5.1-rc1. Of course it was a good idea to not clutter the global
namespaces, but unfortunately it did not turn out to be possible with C
libraries.

See https://sourceware.org/ml/libc-help/2019-03/msg00014.html

Fixes #15775