[cmake] partly revert #15642 to have wayland and GBM working
With the include of Platform.cmake to have for addons is the
GBM_RENDER_SYSTEM not defined, this becomes set from
"./tools/depends/target/Toolchain.cmake" but on a other place.

This remove the change, but after the release must be the cmake
workflow cleaned up. There are some hen and egg problems and
maybe some hidden faults where a value becomes checked which
becomes never available.