[addons] add settings support for property condition (basic conditions only)
Before was it now possible to get a for Kodi defined condition on addon
settings. With this are the basic conditions also available for addon,
to define OS related parts.

The more complex conditions are not added, not sure if this goes then
to far to allow addons also this.