[addons] fix inclusion of cpluff.h
The proper usage of cpluff API is to use #include <cpluff.h> instead of
a hardcoded internal path. The relevant header uses generated files.
This extra header is located in an include path that is already set by
FindCpluff.cmake, so use the main header file from the same location.

Signed-off-by: Olaf Hering <olaf@aepfle.de>