[addons] add possibility to get ID3D11DeviceContext all the time
Have seen by help of afedchin that the context can be change on the fly.
Before was to Screensaver and Visualization this only given during start.

This adds a new addon interface function (`void* kodi::gui::GetHWContext()`)
where the addon can ask about them if needed.

Further is the use of ifdef's removed and becomes handled by WinSystem.h.