[addons] this adds helper functions to get current time
The code happens on "p8platform", but since this defines "long"
as value, it is useless because it is too inaccurate in frames.

Since a lot of addons (over 20) need this I think it belongs to the dev kit.

The code is completely independent of Kodi and gives no access to it.