[addons] add OpenGL support on addon interface / allow GL via DirectX
The main part is the GL.h, to have the integration of the necessary header
OS independently of the addon. Depending on the OS and CMake definition,
this will provide the necessary includes.

With "GLonDX.h" a background class is added which allows to use GL also
under Windows with DirectX.
This is meant for the case that an addon can not easily be converted to
DirectX. If CMake defines a HAS_ANGLE this will be used. In addition,
however, a special "depend" is needed and a customized "FindOpenGLES.cmake".
The support is included in screensaver, visualization and CRendering.