[addons] adds support for various description languages
The fact that the list of addons is stored in the manager, it is not easy
to get the appropriate language. As the language is not yet known when
reading the Replace list and the list would not be correct when changing
the language.

As in the past, the database will be used there, but unfortunately, after
a language change, the database text will continue to be incorrect until
it has been updated again.