Extend list of possible startup windows with Favourites
Having the Favourites dialog in the possible startup window targets is
very convenient as this is where all the important stuff (from different
add-ons live together).

Also nice would be if we could go into the add-on view as well.

Even better would be to add add-on's or register add-on paths to this
list so we could offer a more complete experience.