VideoPlayer: Properly propagate playback errors
Since quite some time, we do not show the playback error dialog
box anymore, when we fail to create a demuxer.

As the m_bAbortRequest-Flag has precedence over the m_error flag,
when we choose between OnPlaybackStopped, OnPlaybackError or
OnPlaybackEnded, we ended up never reporting playback errors.

To fix this issue, we get rid of the double function that
m_bAbortRequest serves, by introducing a m_bCloseRequest-Flag.
This flag is set in CVideoPlayer::CloseFile and used
instead of the m_bAbortRequest-Flag, when we chose callbacks.