[cmake] add support for install bigger addon dir structure
This is an independent change needed on the web browser.

Background is the support of several things at the addon install.

Once the `* _ADDITIONAL_BINARY_EXE` with type 'PROGRAMS' to have the correct settings after the install and to run an EXE is added.

The other is the addition of `* _ADDITIONAL_BINARY_PARTS` and` * _ADDITIONAL_BINARY_DIRS`, which brings a large amount of data to a web browser when you install it.

As addon besides web, a changed inputstream.adaptive was tested with on ZIP install and normal install.