[addons] add interface helper class to translate "C" to "CPP"
This is currently not used, thats why also no Version change is included.
On coming changes with Game and PVR addon system becomes this necessary.

Separately needed at two places and there to keep smaller requests. It
is also usable by others, regardless of the addon types and in the future.