[addons][vis][gui] use IsDirty addon call again
Before was the IsDirty never called as it was alone in a unused Function.

This shifts it to the "Process" call and if "IsDirty" returns true the "MarkDirtyRegion" is called. Most addons do not use the IsDirty and therefore is always true as default, so no big changes occur.

However, if needed now, the IsDirty can also be used on the addon.