[addons][vis] add function to transfer presets and increase version to 2.0.2
This is unfortunately necessary because of a small hen and egg problem. When using Angle to Windows, the GL is only supported at the startup call.

To ProjectM e.g. the presets can only be captured after the `new projectM` call, which in turn requires GL.
Hereby it is possible to transmit this during the "Start" call from the addon.

The new clear function is only used within the headers to avoid flooding on multiple calls.