[settings] allow set of needed extensions for picture selection
Before was always all for Kodi available picture extensions present.
This allow to specialize the needed one.

Until now is only over normal format "file" the set of needed extensions
possible, but there on new settings version not possible to get a icon
of wanted files. On old was this done with special ways, where the backport
support set this.

But on newer settings.xml was it no more possible where `format="image"` is
needed, this change now also on new the setting about.

e.g.
```xml
        <setting id="texture0" type="path" label="30016" help="30017">
          <default></default>
          <dependencies>
            <dependency type="enable" setting="ownshader" operator="is">true</dependency>
            <dependency type="enable" setting="texture0-sound" operator="is">false</dependency>
          </dependencies>
          <constraints>
            <allowempty>true</allowempty>
            <masking>*.png</masking> <!-- HERE now possible to set for image -->
          </constraints>
          <control type="button" format="image">
            <heading>30016</heading>
          </control>
        </setting>
```