Make tag processing of "various artists" compilation albums more consistent and less impacted by changes in language settings.
Gather such compilations under one unique artist regardless of whether they are tagged with album artist of "various", "various artists", a localized equivalent (in current language), with the VA mbid (and whatever name) or if no album artist is specified but the song artists are mixed. Do not attempt to translate artist name from the tag value, simply add the artist the first time it is encountered using the tag value and unique mbid in the same way any other artist would be added.