[skin.estuary] Improvement: 'Favourites' SideBlade Action
Description
Favourites SideBlade action addition.

Motivation and Context
I thought the SideBlade menu could use some more functionality. In my vision it is an always accessible sidemenu that could include shortcuts. Added a "Favourites" shortcut above the "Get more..." option. I will not be disappointed if it does not get included, but it is here to serve as food for thought for more sideblade actions.

How Has This Been Tested?
Tested on Android and Windows using latest stable Leia.

Screenshots (if appropriate):
![Screenshot_20191210-034121](https://user-images.githubusercontent.com/43998058/70494978-eea0cc00-1b15-11ea-898a-a53036ad223c.jpg)


Types of change
[x] Improvement (non-breaking change which improves existing functionality)

Checklist:
[x]  My code follows the Code Guidelines of this project