Windowing: GBM - Fix refreshrate switch
We wake up the audio engine way too early.
In order to do it correctly, we need to wait until
the modeset has been executed by the atomic commit.
Otherwise there is a big chance that the audio engine
is trying to open the sound device while the ELD info
is broken.