[cmake] fix not allowed depends check
Before was in case with more as one value for not allowed OS
e.g. "!windows !windowsstore" defined in platforms.txt, it used
all the time.

Is a cmake "for" loop where it compare the name with wanted target,
before has it seen the name is, in one of the names not equal and
set to allowed and continued loop, no matter if on next/before place defined
as not.

With this change it check the not allowed name is present,
make `set(${build} FALSE)` and break the loop.

Related to inputstream.ffmpegdirect, where on some depends, Windows and
Windowsstore must be prevented.