[addons] use on CAddonDll instance class pointer as identifier
Before was a string used (where was in most cases only a pointer as string).
Seen during PVR rework, where it becomes needed to use a own identifier, to pass
same name several times to identify e.g. "channel" or "recording".

With same string then comes in conflict with the std::map where store used
Instances on CAddonDll, by use of the related class pointer this is more safe.

Only was related to game addon system, where currently not use
IAddonInstanceHandler (games to complex for me to find direct a way with them),
a add of ADDON_INSTANCE_HANDLER (const void*) needed.

On addon interface change this nothing, only store the map in Kodi it by
different value.