[addons] move binary addon interface calls to own files
Previously, these functions were within CAddonDll, but since its function
is primarily the basis for Kodi of the addon, but not its interface, these
functions did not fit there.

Since these form the API, they belong in the associated interface folder.
This allows a much better overview, and the name "AddonBase.*"
used there corresponds to the associated addon header.