[addons] fix stupid circumstance with addon settings on folder with file extension
Found this when adding settings to vfs.rar.

There called this place with "special: //profile/addon_data/vfs.rar", this was seen
by "CFileDirectoryFactory::Create" as a file and tried to open with vfs.rar addon.

This logically leads to an error because it is only a folder.

The slash is left now on the folder so that it is recognized as a folder.