Add section on Meson and nasm build dependency in Ubuntu
In order to build dav1d (used as AV1 decoder in ffmpeg),
an up-to-date Meson and nasm package is required. dav1d
currently requires at least meson >= 0.47.0 and nasm >= 2.14:
https://code.videolan.org/videolan/dav1d/-/blob/0.6.0/meson.build#L31

Since Ubuntu 18.04 ships with meson 0.45.1 and nasm 2.13.02,
Meson and nasm have to be updated manually.

Support for AV1 was added in #16933.

Relates to https://github.com/xbmc/xbmc/pull/16933#issuecomment-597366773.