[addons][videocodec] change buffer handle to use "KODI_HANDLE"
Before was this used as "void *buffer" and thought first that is
something used from addon itself and then seen it is "CVideoBuffer*"
class from Kodi and given to addon for interact back to Kodi.

This change it to "KODI_HANDLE videoBufferHandle" on Kodi's calls, but
leaved on VIDEOCODEC_PICTURE by buffer only to prevent compatibility
problems. This then done on major changes.

With KODI_HANDLE is more easy to see that something in Kodi itself related.