[addons][videocodec][demux] add 'extern "C"' to DemuxPacket
This is a minor change taken from another request to reduce his size.
On DemuxPacket structure is nothing changed only is the extern "C" added
and related to clang 2 characters moved forward.

The DemuxPacket.h need to change on other request to have only "C" in
structure.

As Future Info: For that becomes then a Kodi side class added where has this
as parent and his constructor parts removed there.