[addons][vfs] do not interact addon class with Kodi
This was before given to Kodi to have on his calls to addon.
That basically not match a "C" ABI where the use of "C++" parts within
"C" not allowed!

With this it becomes given via "void*" to Kodi and with a "static_cast"
changed back to required class on addon.