[settings] allow TranslatableIntegerSettingOption to get string from addon
Before was this pair with "integer, integer" value, where first string
identifier and second setting value identifier. But with this was it
not possible to handle them on addons where can give own strings.

This change it from std::pair to a structure like the others already have
and with add of third value to get in case of addon string his related
addon identifier.

For this it looks a bit bigger as some places are touched to change from
pair to class constructor.