[addons][pvr] Pass in addon string ID and not its class
This change is an independent part of the rework to switch to the new API.
This removes the use of CAddonDll in some places and gives its string
identification instead.

This is not critical since the called up positions only need this.

The background to this change is that in the multiple instance way the
addon class is no longer usable because it can also affect other instances.

The other reason is to reduce the rework request in size.