[addons][pvr] add GetPlayingChannelUniqueID to playback state
This add "int CPVRPlaybackState::GetPlayingChannelUniqueID() const" to get
on wanted places the currently running PVR channel unique identifier.

This are also now used by "CPVRGUIInfo::UpdateQualityData()" and
"CPVRGUIInfo::UpdateDescrambleData()" and this !IsPlayingRecording()
becomes replaced with them.

GetPlayingChannelUniqueID() match more the wanted target as to ask
that is not a recording.

Further is this value also needed on reworked PVR system to pass to addon.
But this change here is mostly independent and thats why separated