[addons][vfs] fix Kodi start crash if incompatible VFS addon present
Before does Kodi always crash if a incompatible VFS addon was present, that
was comming during his version checks on start and try to send a GUI message
without a available GUI!

This add a check about GUIManager is present and prevent his call if not.
The wanted message is still be shown as this is cached and handled after
GUI start.

Backtrace:
```
Thread 1 "kodi-x11" received signal SIGSEGV, Segmentation fault.
0x00005555565a00c4 in CGUIComponent::GetWindowManager() ()
(gdb) bt
```